標準入出力にコマンドを追加
