+++ src/os_v5/system_calls/interface.ts
@@ -6,7 +6,7 @@ import { EnvironmentVariable } from "./environment_variable"
 import { UniqueId } from "./unique_id"
 import { UniqueName } from "./game_system_calls/unique_name"
 // import { StartupLauncher } from "./depended_system_calls/startup_launcher"
-import { StandardIO } from "./depended_system_calls/standard_io"
+import { StandardIO } from "./standard_io/standard_io"
 
 class SystemCallList {
   // Primitive
