+++ src/os_v5/system_calls/standard_io/command.ts
@@ -1,0 +1,1 @@
+export type Command = (args: string[]) => string
